+++ extensions/cornerstone-dicom-seg/src/getPanelModule.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 
 import { useAppConfig } from '@state';
-import { Toolbox } from '@ohif/ui';
+import { Toolbox as NewToolbox } from '@ohif/ui-next';
+import { Toolbox as OldToolbox } from '@ohif/ui';
 import PanelSegmentation from './panels/PanelSegmentation';
 
 const getPanelModule = ({
@@ -12,7 +13,28 @@ const getPanelModule = ({
   title,
 }: withAppTypes) => {
   const { customizationService } = servicesManager.services;
-  const wrappedPanelSegmentation = configuration => {
+
+  const wrappedPanelSegmentation = ({ configuration, renderHeader, getCloseIcon, tab }) => {
+    const [appConfig] = useAppConfig();
+
+    return (
+      <PanelSegmentation
+        commandsManager={commandsManager}
+        servicesManager={servicesManager}
+        extensionManager={extensionManager}
+        configuration={{
+          ...configuration,
+          disableEditing: appConfig.disableEditing,
+          ...customizationService.get('segmentation.panel'),
+        }}
+        renderHeader={renderHeader}
+        getCloseIcon={getCloseIcon}
+        tab={tab}
+      />
+    );
+  };
+
+  const wrappedPanelSegmentationNoHeader = ({ configuration }) => {
     const [appConfig] = useAppConfig();
     const modeCustomization = customizationService.getModeCustomization('segmentation.panel');
 
@@ -30,9 +52,15 @@ const getPanelModule = ({
     );
   };
 
-  const wrappedPanelSegmentationWithTools = configuration => {
+  const wrappedPanelSegmentationWithTools = ({
+    configuration,
+    renderHeader,
+    getCloseIcon,
+    tab,
+  }) => {
     const [appConfig] = useAppConfig();
-    const modeCustomization = customizationService.getModeCustomization('segmentation.panel');
+
+    const Toolbox = appConfig.useExperimentalUI ? NewToolbox : OldToolbox;
 
     return (
       <>
@@ -45,6 +73,9 @@ const getPanelModule = ({
           configuration={{
             ...configuration,
           }}
+          renderHeader={renderHeader}
+          getCloseIcon={getCloseIcon}
+          tab={tab}
         />
         <PanelSegmentation
           commandsManager={commandsManager}
@@ -74,6 +105,13 @@ const getPanelModule = ({
       iconLabel: 'Segmentation',
       label: 'Segmentation',
       component: wrappedPanelSegmentationWithTools,
+    },
+    {
+      name: 'panelSegmentationNoHeader',
+      iconName: 'tab-segmentation',
+      iconLabel: 'Segmentation',
+      label: 'Segmentation',
+      component: wrappedPanelSegmentationNoHeader,
     },
   ];
 };