+++ extensions/default/src/Components/SidePanelWithServices.tsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useState, useCallback } from 'react';
-import { SidePanel } from '@ohif/ui';
+import { SidePanel as NewSidePanel } from '@ohif/ui-next';
+import { SidePanel as OldSidePanel } from '@ohif/ui';
+import { useAppConfig } from '@state';
 import { Types } from '@ohif/core';
 
 export type SidePanelWithServicesProps = {
@@ -26,6 +28,7 @@ const SidePanelWithServices = ({
   const [hasBeenOpened, setHasBeenOpened] = useState(false);
   const [activeTabIndex, setActiveTabIndex] = useState(activeTabIndexProp);
   const [tabs, setTabs] = useState(tabsProp ?? panelService.getPanels(side));
+  const [appConfig] = useAppConfig();
 
   const handleSidePanelOpen = useCallback(() => {
     setHasBeenOpened(true);
@@ -74,6 +77,8 @@ const SidePanelWithServices = ({
       activatePanelSubscription.unsubscribe();
     };
   }, [tabs, hasBeenOpened, panelService]);
+
+  const SidePanel = appConfig?.useExperimentalUI ? NewSidePanel : OldSidePanel;
 
   return (
     <SidePanel