+++ extensions/default/src/Panels/index.js
@@ -1,4 +1,4 @@
-import PanelStudyBrowser from './PanelStudyBrowser';
+import PanelStudyBrowser from './StudyBrowser/PanelStudyBrowser';
 import WrappedPanelStudyBrowser from './WrappedPanelStudyBrowser';
 import PanelMeasurementTable from './PanelMeasurementTable';
 import createReportDialogPrompt from './createReportDialogPrompt';