+++ platform/ui-next/src/components/index.ts
@@ -6,15 +6,28 @@ import Combobox from './Combobox';
 import Popover from './Popover';
 import Calendar from './Calendar';
 import DatePickerWithRange from './DateRange';
+import Separator from './Separator';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from './Tabs';
+import { Toggle, toggleVariants } from './Toggle';
+import { ToggleGroup, ToggleGroupItem } from './ToggleGroup';
 import { Input } from './Input';
 import { Label } from './Label';
-import { Tabs, TabsList, TabsTrigger, TabsContent } from './Tabs';
-import { Separator } from './Separator';
 import { Switch } from './Switch';
 import { Checkbox } from './Checkbox';
-import { Toggle, toggleVariants } from './Toggle';
 import { Slider } from './Slider';
 import { ScrollArea, ScrollBar } from './ScrollArea';
+import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from './Accordion';
+import { Icons } from './Icons';
+import SidePanel from './SidePanel';
+import StudyItem from './StudyItem';
+import StudyBrowser from './StudyBrowser';
+import StudyBrowserSort from './StudyBrowserSort';
+import StudyBrowserViewOptions from './StudyBrowserViewOptions';
+import Thumbnail from './Thumbnail';
+import ThumbnailList from './ThumbnailList';
+import PanelSection from './PanelSection';
+import DisplaySetMessageListTooltip from './DisplaySetMessageListTooltip';
+import { Toolbox, ToolboxUI } from './Toolbox';
 
 export {
   Button,
@@ -39,4 +52,23 @@ export {
   toggleVariants,
   Slider,
   ScrollArea,
+  ToggleGroup,
+  ToggleGroupItem,
+  ScrollBar,
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+  Icons,
+  SidePanel,
+  StudyItem,
+  StudyBrowser,
+  StudyBrowserSort,
+  StudyBrowserViewOptions,
+  Thumbnail,
+  ThumbnailList,
+  PanelSection,
+  DisplaySetMessageListTooltip,
+  Toolbox,
+  ToolboxUI,
 };